+++ BetterCrafting/Patches/LetterViewerMenu_Patches.cs
@@ -1,14 +1,10 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 using HarmonyLib;
+
 using StardewModdingAPI;
 
 using StardewValley;
-using StardewValley.GameData.Movies;
 using StardewValley.Menus;
 
 namespace Leclair.Stardew.BetterCrafting.Patches;
@@ -34,7 +30,7 @@ public static class LetterViewerMenu_Patches {
 	public static void HandleItemCommand_Prefix(LetterViewerMenu __instance, ref string mail) {
 		try {
 			mail = ProcessMail(__instance, mail);
-		} catch(Exception ex) {
+		} catch (Exception ex) {
 			Monitor?.Log($"An error occurred while attempting to process item commands in a LetterViewerMenu.", LogLevel.Error);
 			Monitor?.Log($"Details:\n{ex}", LogLevel.Error);
 		}
@@ -45,7 +41,7 @@ public static class LetterViewerMenu_Patches {
 		ModEntry mod = ModEntry.Instance;
 
 		int searchFromIndex = 0;
-		while ( true ) {
+		while (true) {
 			int startItemIndex = input.IndexOf("%item", searchFromIndex, StringComparison.InvariantCulture);
 			if (startItemIndex < 0)
 				break;
@@ -59,14 +55,14 @@ public static class LetterViewerMenu_Patches {
 			string type = typeAndArgs[0];
 			string[] args = typeAndArgs.Length > 1
 				? ArgUtility.SplitBySpace(typeAndArgs[1])
-				: Array.Empty<string>();
+				: [];
 
 			bool doReplace = false;
 
 			// Do we have a craftingrecipe or cookingrecipe?
-			if (! menu.isFromCollection && (type.ToLower() == "cookingrecipe" || type.ToLower() == "craftingrecipe")) {
+			if (!menu.isFromCollection && (type.ToLower() == "cookingrecipe" || type.ToLower() == "craftingrecipe")) {
 				string recipeKey = string.Join(" ", args);
-				bool cooking = type.ToLower() == "cookingrecipe";
+				//bool cooking = type.ToLower() == "cookingrecipe";
 
 				if (!string.IsNullOrEmpty(recipeKey)) {
 					if (mod.DataRecipes.TryGetRecipeById(recipeKey, out var recipe)) {